refactor(orch2): split main.py into focused modules
Per PR review:

- pool.py:            Pool, EnvWorker, cycle_forever, build_pool
- inference_admin.py: InferenceAdmin
- batcher.py:         TrainBatcher, Done (the max_steps sentinel)
- engine.py:          Group, Inflight, RolloutEngine (Pool now imported)
- main.py:            glue only — asserts, logging, wiring, TaskGroup

No behavior change, just file layout.
